Test `where` on mismatched types
In order to make sure type promotion occurs correctly, have chosen the
two arrays to have different types. Should be promoted to the largest of
the two types.

+++ dask/array/tests/test_array_core.py
@@ -661,7 +661,7 @@ def test_choose():
 def test_where():
     x = np.random.randint(10, size=(15, 16))
     d = from_array(x, chunks=(4, 5))
-    y = np.random.randint(10, size=15)
+    y = np.random.randint(10, size=15, dtype=np.uint8)
     e = from_array(y, chunks=(4,))
 
     assert_eq(where(d > 5, d, 0), np.where(x > 5, x, 0))
